requirements.txt
streamlit>=1.55,<2
pandas>=2.3,<3
numpy>=2.4,<3
yfinance>=1.2,<2
feedparser>=6.0,<7
python-dateutil>=2.9,<3
python-dotenv>=1.0,<2
requests>=2.32,<3
